init match module
