+++ composeApp/build.gradle.kts
@@ -46,9 +46,10 @@ kotlin {
             implementation(project(":core:network"))
             implementation(project(":core:database"))
             implementation(project(":core:ui"))
-            implementation(project(":feature:feed"))
+            implementation(project(":feature:film_feed"))
             implementation(project(":feature:film"))
             implementation(project(":feature:profile"))
+            implementation(project(":feature:match_feed"))
         }
     }
 }
