+++ composeApp/src/commonMain/kotlin/main_screen/bottom_nav_bar/FeedTab.kt
@@ -9,7 +9,7 @@ import cafe.adriel.voyager.navigator.Navigator
 import cafe.adriel.voyager.navigator.tab.Tab
 import cafe.adriel.voyager.navigator.tab.TabOptions
 import cafe.adriel.voyager.transitions.SlideTransition
-import com.stslex.feature.feed.ui.FeedScreen
+import com.stslex.feature.film_feed.ui.FeedScreen
 
 object FeedTab : Tab {
