--- feature/feed/src/commonMain/kotlin/com/stslex/feature/feed/data/model/FeedDataMapper.kt
@@ -1,18 +1,0 @@
-package com.stslex.feature.feed.data.model
-
-import com.stslex.core.network.clients.film.model.FilmFeedResponse
-import com.stslex.core.network.clients.film.model.FilmResponse
-
-fun FilmFeedResponse.toData() = FeedDataModel(
-    films = results.map { it.toData() },
-    hasNextPage = hasNext,
-)
-
-fun FilmResponse.toData() = FilmDataModel(
-    id = id,
-    title = title,
-    description = description,
-    imageUrl = poster,
-    rate = rating,
-    genres = genres,
-)
