--- feature/feed/src/commonMain/kotlin/com/stslex/feature/feed/data/model/FeedDataModel.kt
@@ -1,6 +1,0 @@
-package com.stslex.feature.feed.data.model
-
-data class FeedDataModel(
-    val films: List<FilmDataModel>,
-    val hasNextPage: Boolean,
-)
