--- feature/feed/src/commonMain/kotlin/com/stslex/feature/feed/data/repository/FeedRepository.kt
@@ -1,8 +1,0 @@
-package com.stslex.feature.feed.data.repository
-
-import com.stslex.feature.feed.data.model.FeedDataModel
-
-interface FeedRepository {
-
-    suspend fun getFeed(page: Int, pageSize: Int): FeedDataModel
-}
