--- feature/feed/src/commonMain/kotlin/com/stslex/feature/feed/domain/interactor/FeedInteractor.kt
@@ -1,8 +1,0 @@
-package com.stslex.feature.feed.domain.interactor
-
-import com.stslex.feature.feed.domain.model.FeedDomainModel
-
-interface FeedInteractor {
-
-    suspend fun getFeed(page: Int, pageSize: Int): FeedDomainModel
-}
