--- feature/feed/src/commonMain/kotlin/com/stslex/feature/feed/domain/model/FeedDomainModel.kt
@@ -1,6 +1,0 @@
-package com.stslex.feature.feed.domain.model
-
-data class FeedDomainModel(
-    val films: List<FilmDomainModel>,
-    val hasNextPage: Boolean,
-)
