--- feature/feed/src/commonMain/kotlin/com/stslex/feature/feed/domain/model/FilmDomainModel.kt
@@ -1,10 +1,0 @@
-package com.stslex.feature.feed.domain.model
-
-data class FilmDomainModel(
-    val id: String,
-    val title: String,
-    val description: String,
-    val imageUrl: String,
-    val rate: String,
-    val genres: List<String>,
-)
