--- feature/feed/src/commonMain/kotlin/com/stslex/feature/feed/navigation/FeedScreenRouter.kt
@@ -1,6 +1,0 @@
-package com.stslex.feature.feed.navigation
-
-import com.stslex.core.ui.navigation.Router
-import com.stslex.feature.feed.ui.store.FeedScreenStoreComponent.Navigation
-
-interface FeedScreenRouter : Router<Navigation>
