--- feature/feed/src/commonMain/kotlin/com/stslex/feature/feed/ui/components/FeedScreenLoading.kt
@@ -1,21 +1,0 @@
-package com.stslex.feature.feed.ui.components
-
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material3.CircularProgressIndicator
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-
-@Composable
-internal fun FeedScreenLoading(
-    modifier: Modifier = Modifier
-) {
-    // TODO add loading screen
-    Box(
-        modifier = modifier.fillMaxSize(),
-        contentAlignment = Alignment.Center
-    ) {
-        CircularProgressIndicator()
-    }
-}
