+++ feature/film_feed/src/commonMain/kotlin/com/stslex/feature/film_feed/data/model/FeedDataModel.kt
@@ -1,0 +1,6 @@
+package com.stslex.feature.film_feed.data.model
+
+data class FeedDataModel(
+    val films: List<FilmDataModel>,
+    val hasNextPage: Boolean,
+)
