+++ feature/film_feed/src/commonMain/kotlin/com/stslex/feature/film_feed/data/model/FilmDataModel.kt
@@ -1,0 +1,10 @@
+package com.stslex.feature.film_feed.data.model
+
+data class FilmDataModel(
+    val id: String,
+    val title: String,
+    val description: String,
+    val imageUrl: String,
+    val rate: String,
+    val genres: List<String>,
+)
