+++ feature/film_feed/src/commonMain/kotlin/com/stslex/feature/film_feed/data/repository/FeedRepository.kt
@@ -1,0 +1,8 @@
+package com.stslex.feature.film_feed.data.repository
+
+import com.stslex.feature.film_feed.data.model.FeedDataModel
+
+interface FeedRepository {
+
+    suspend fun getFeed(page: Int, pageSize: Int): FeedDataModel
+}
