+++ feature/film_feed/src/commonMain/kotlin/com/stslex/feature/film_feed/data/repository/FeedRepositoryImpl.kt
@@ -1,0 +1,21 @@
+package com.stslex.feature.film_feed.data.repository
+
+import com.stslex.core.network.clients.film.client.FilmClient
+import com.stslex.feature.film_feed.data.model.FeedDataModel
+import com.stslex.feature.film_feed.data.model.toData
+
+class FeedRepositoryImpl(
+    private val client: FilmClient
+) : FeedRepository {
+
+    override suspend fun getFeed(
+        page: Int,
+        pageSize: Int
+    ): FeedDataModel = client
+        .getFeedFilms(
+            page = page,
+            pageSize = pageSize
+        )
+        .toData()
+}
+
