+++ feature/film_feed/src/commonMain/kotlin/com/stslex/feature/film_feed/domain/interactor/FeedInteractor.kt
@@ -1,0 +1,8 @@
+package com.stslex.feature.film_feed.domain.interactor
+
+import com.stslex.feature.film_feed.domain.model.FeedDomainModel
+
+interface FeedInteractor {
+
+    suspend fun getFeed(page: Int, pageSize: Int): FeedDomainModel
+}
