+++ feature/film_feed/src/commonMain/kotlin/com/stslex/feature/film_feed/domain/model/FeedDomainModel.kt
@@ -1,0 +1,6 @@
+package com.stslex.feature.film_feed.domain.model
+
+data class FeedDomainModel(
+    val films: List<FilmDomainModel>,
+    val hasNextPage: Boolean,
+)
