+++ feature/film_feed/src/commonMain/kotlin/com/stslex/feature/film_feed/navigation/FeedScreenRouter.kt
@@ -1,0 +1,6 @@
+package com.stslex.feature.film_feed.navigation
+
+import com.stslex.core.ui.navigation.Router
+import com.stslex.feature.film_feed.ui.store.FeedScreenStoreComponent.Navigation
+
+interface FeedScreenRouter : Router<Navigation>
