+++ feature/film_feed/src/commonMain/kotlin/com/stslex/feature/film_feed/navigation/FeedScreenRouterImpl.kt
@@ -1,0 +1,20 @@
+package com.stslex.feature.film_feed.navigation
+
+import com.stslex.core.ui.navigation.AppNavigator
+import com.stslex.core.ui.navigation.AppScreen
+import com.stslex.feature.film_feed.ui.store.FeedScreenStoreComponent
+
+class FeedScreenRouterImpl(
+    private val navigator: AppNavigator
+) : FeedScreenRouter {
+
+    override fun invoke(event: FeedScreenStoreComponent.Navigation) {
+        when (event) {
+            is FeedScreenStoreComponent.Navigation.Film -> navigateToFilm(event)
+        }
+    }
+
+    private fun navigateToFilm(event: FeedScreenStoreComponent.Navigation.Film) {
+        navigator.navigate(AppScreen.Film(event.filmId))
+    }
+}
