+++ feature/film_feed/src/commonMain/kotlin/com/stslex/feature/film_feed/ui/components/FeedScreenError.kt
@@ -1,0 +1,21 @@
+package com.stslex.feature.film_feed.ui.components
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+
+@Composable
+internal fun FeedScreenError(
+    message: String,
+    modifier: Modifier = Modifier
+) {
+    // TODO add error screen
+    Box(
+        modifier = modifier,
+        contentAlignment = Alignment.Center
+    ) {
+        Text(text = message)
+    }
+}
