+++ feature/match_feed/src/commonMain/kotlin/com/stslex/feature/match_feed/MatchFeedScreen.kt
@@ -1,0 +1,3 @@
+package com.stslex.feature.match_feed
+
+object MatchFeedScreen
