+++ iosApp/FeatureFilmFeedPodfile
@@ -1,0 +1,5 @@
+target 'test' do
+  use_frameworks!
+  platform :ios, '16.0'
+  pod 'film_feed', :path => '../feature/film_feed'
+end
