+++ iosApp/FeatureMatchFeedPodfile
@@ -1,0 +1,5 @@
+target 'test' do
+  use_frameworks!
+  platform :ios, '16.0'
+  pod 'match_feed', :path => '../feature/match_feed'
+end
