+++ settings.gradle.kts
@@ -24,6 +24,7 @@ include(":core:core")
 include(":core:network")
 include(":core:ui")
 include(":core:database")
-include(":feature:feed")
+include(":feature:film_feed")
 include(":feature:film")
 include(":feature:profile")
+include(":feature:match_feed")
